refactor: update import statement in test files
